+++ resources/js/Layouts/GuestLayout.tsx
@@ -1,4 +1,4 @@
-import ApplicationLogo from "@/Components-template/ApplicationLogo";
+import ApplicationLogo from "@/Components-breeze/ApplicationLogo";
 import { Link } from "@inertiajs/react";
 import { PropsWithChildren } from "react";
