+++ resources/js/Pages/Auth/VerifyEmail.tsx
@@ -1,4 +1,4 @@
-import PrimaryButton from "@/Components-template/PrimaryButton";
+import PrimaryButton from "@/Components-breeze/PrimaryButton";
 import GuestLayout from "@/Layouts/GuestLayout";
 import { Head, Link, useForm } from "@inertiajs/react";
 import { FormEventHandler } from "react";
